backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pillow==11.1.0
transformers==4.49.0
torch>=2.6,<2.11
ultralytics==8.3.78
sentencepiece>=0.2,<0.3
